refactor: move Navigation to Components
